fix(server,openomni): register plan agent and deduplicate memory adapter
