+++ apps/server/src/agents/registry.ts
@@ -1,8 +1,11 @@
 import type { AgentFactory, AgentDefinition, AgentPromptMetadata } from "./types";
 import { createDevAgent, devAgentMeta } from "./dev-agent/index";
-import { planAgentMeta } from "./plan-agent/index";
+import { createPlanAgent, planAgentMeta } from "./plan-agent/index";
 
-const agentSources = new Map<string, AgentFactory>([["dev", createDevAgent]]);
+const agentSources = new Map<string, AgentFactory>([
+  ["dev", createDevAgent],
+  ["plan", createPlanAgent],
+]);
 const metadata = new Map<string, AgentPromptMetadata>([
   ["dev", devAgentMeta],
   ["plan", planAgentMeta],
